requirements.txt
gradio>=4.0.0
langchain>=0.1.0
langchain-community>=0.0.20
faiss-cpu>=1.7.4
numpy>=1.24.0
pydantic>=2.0.0
python-dotenv>=1.0.0
sentence-transformers>=2.2.0
chromadb>=0.4.0
requests>=2.31.0
markdown>=3.5.0
huggingface-hub>=0.22.0
transformers>=4.44.0